% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a10.R
\name{a10}
\alias{a10}
\title{Main a10 function}
\usage{
a10(nf_init, nm_init, sexdeb_init_f, sexdeb_init_m, init_prev_f, init_prev_m,
  beta_m2f, beta_f2m, pc_debuting_f, pc_debuting_m, coital_acts_pp_f,
  coital_acts_pp_m, condom_use_f, condom_use_m, ann_chg_npartners,
  ann_chg_coital, ann_chg_condoms)
}
\arguments{
\item{x}{A vector}
}
\value{
A matrix with 3 rows containing the values from x filled in by row
}
\description{
Main a10 function
}
