% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fx.R
\name{array11}
\alias{array11}
\title{Make an array that repeats a matrix 11 times}
\usage{
array11(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A matrix with 11 laters with each layer equaling x
}
\description{
Make an array that repeats a matrix 11 times
}
