% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_inschool_calc.R
\name{diag_inschool_calc}
\alias{diag_inschool_calc}
\title{Function to convert all diagnoses to diagnoses in school 
 NA: to be run separately by sex}
\usage{
diag_inschool_calc(diags, perc_diag, infdur)
}
\arguments{
\item{XX}{A 3x6 matrix of past-year diagnoses among}

\item{XX}{A 3x6 matrix of precent of cases that get diagnosed}

\item{XX}{A 3x6 matrix indicating mean dur of infection}
}
\value{
A 3 x 6 matrix indicating est. number of prevalent cases in the cross-section
}
\description{
Function to convert all diagnoses to diagnoses in school 
 NA: to be run separately by sex
}
