% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPrev_calc.R
\name{initPrev_calc}
\alias{initPrev_calc}
\title{Function calculate initial prevalence 
 NA: to be run separately forby sex}
\usage{
initPrev_calc(diags, perc_diag, infdur)
}
\arguments{
\item{diags}{A 3x6 matrix of past-year diagnoses among}

\item{perc_diag}{A 3x6 matrix of precent of cases that get diagnosed}

\item{infdur}{A 3x6 matrix indicating mean dur of infection}
}
\value{
A 3 x 6 matrix indicating est. number of prevalent cases in the cross-section
}
\description{
Function calculate initial prevalence 
 NA: to be run separately forby sex
}
