% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fx.R
\name{mat2}
\alias{mat2}
\title{Make a matrix with 2 rows from a vector}
\usage{
mat2(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A matrix with 2 rows containing the values from x filled in by row
}
\description{
Make a matrix with 2 rows from a vector
}
