% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fx.R
\name{mat3}
\alias{mat3}
\title{Make a matrix with 3 rows from a vector}
\usage{
mat3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A matrix with 3 rows containing the values from x filled in by row
}
\description{
Make a matrix with 3 rows from a vector
}
