% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_inschool_calc.R
\name{diag_inschool_calc}
\alias{diag_inschool_calc}
\title{Function to convert all diagnoses to diagnoses in school 
 NA: to be run separately by sex}
\usage{
diag_inschool_calc(diags, perc_inschool)
}
\arguments{
\item{diags}{A 3x6 matrix of past-year diagnoses among adols regardless of school status}

\item{perc_inschool}{A 3x6 matrix of precent of adols in HS}
}
\value{
A 3x6 matrix indicating number of diagnoses among HS students
}
\description{
Function to convert all diagnoses to diagnoses in school 
 NA: to be run separately by sex
}
